Crazy new way to do virtualizing in lists, need to react to new items and resize events
