+++ src/BlazorFluentUI.BFUList/BFUListOLD.razor
@@ -1,4 +1,4 @@
-﻿@namespace BlazorFluentUI
+﻿@*@namespace BlazorFluentUI
 @inherits BFUComponentBase
 
 @typeparam TItem
@@ -10,4 +10,4 @@
     <div @ref="surfaceDiv" @onscroll=@OnScroll class="ms-List-surface" role="presentation" style="overflow-y:auto;height:100%;">
         @ItemPagesRender
     </div>
-</div>
+</div>*@
